pyproject.toml
[tool.poetry]
name = "valuagent"
version = "0.1.0"
description = ""
authors = ["Jan Helcl"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.0"
google-genai = "^1.30.0"
python-dotenv = "^1.1.1"
jupyterlab = "^4.4.5"
ipykernel = "^6.30.0"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
python-multipart = "^0.0.9"
openpyxl = "^3.1.5"
slowapi = "^0.1.9"
google-cloud-logging = "^3.10.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
